Revert change

+++ src/zAppDev.DotNet.Framework/Mvc/Helper/MapHelper.cs
@@ -106,8 +106,8 @@ namespace zAppDev.DotNet.Framework.Mvc
 
     public class MapHelperCoordinates
     {
-        public decimal? Latitude { get; set; }
-        public decimal? Longitude { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
     }
 
     public class GeoJsonObject
